+++ docs/releasenotes.md
@@ -3,6 +3,7 @@
 * Optimizations have been made to further improve the overall performance and stability of *DOOM Retro*.
 * The *Fortress Of Mystery* is now shown on the intermission screen for the remainder of the episode when the player finds the secret exit to E2M9 in `doom.wad`.
 * A bug has been fixed whereby some player stats saved in `doomretro.cfg` could become corrupted in some instances.
+* Minor changes have been made to text that is output to the console.
 * A bug has been fixed whereby it was possible for some controls to be bound twice to the same action in `doomretro.cfg`.
 
 ---